Added code for game
While this version of the code is not something i want to show off, hence why I did not include it in past commits, I will be fixing it by following the DRY principle and rewriting the code with a general aim to make it clean and efficient. So there is nothing to worry about as in future commits the code will only get better and better and hopefully the commit description will get shorter and shorter.