Code Revamp One: Added Background, Player and Floor
The previous code was so bad that it is easier to just completely rewrite the code than to make edits to the old code thus this is indeed a 'revamp'. There is no gameplay yet, just a still player and the map. Code takes advantage of object oriented JavaScript and has much better documentation.